+++ tests/testthat/integration_tests/test_infrastructure.R
@@ -1,39 +1,42 @@
 context("infrastructure-non cran tests")
 
+proj <- usethis::proj_get()
+
 test_that("use_testdata creates testdata dir", {
   #* @testing has_testdata
   #* @testing use_testdata
   package_state <- list.files(".", recursive = TRUE)
-
   tpkg <- file.path(rprojroot::find_testthat_root_file("testdata", "test_pkg"))
+  usethis::proj_set(tpkg)
+
 
   # Check for clean state
-  expect_false(has_testdata(tpkg))
+  expect_false(has_testdata())
 
   # Check creation of testdata dir
   expect_output(
     expect_message(
-      use_testdata(iris, base_path = tpkg),
+      use_testdata(iris),
       "You can save data files for tests via"
     ),
     "Creating"
   )
 
   expect_true(dir.exists(file.path(tpkg, "tests/testthat/testdata")))
-  expect_true(has_testdata(base_path = tpkg))
+  expect_true(has_testdata())
 
   # Cleanup
   unlink(
     file.path(tpkg, "tests/testthat/testdata"),
     recursive = TRUE
   )
-  expect_false(has_testdata(tpkg))
+  expect_false(has_testdata())
 
 
   # Check creation of testdata dfiles
   expect_output(
     expect_message(
-      use_testdata(iris, base_path = tpkg),
+      use_testdata(iris),
       "You can save data files for tests via"
     ),
     "Creating"
@@ -41,7 +44,7 @@ test_that("use_testdata creates testdata dir", {
 
   expect_output(
     expect_message(
-      use_testdata(iris, subdir = "iris", base_path = tpkg),
+      use_testdata(iris, subdir = "iris"),
       "You can save data files for tests via"
     ),
     "Creating"
@@ -64,6 +67,7 @@ test_that("use_testdata creates testdata dir", {
   # Check for unexpected changes to package
   package_state_new <- list.files(".", recursive = TRUE)
   expect_identical(package_state, package_state_new)
+  usethis::proj_set(proj)
 })
 
 
@@ -74,10 +78,11 @@ test_that("use_tester works as expected", {
 
   #* @testing use_tester
   tpkg <- rprojroot::find_testthat_root_file("testdata", "test_pkg")
+  usethis::proj_set(tpkg)
 
   # Check if tester file is created at the correct path and not empty
   expect_message(expect_true(
-    use_tester("footests", base_path = tpkg, ignore = FALSE)),
+    use_tester("footests", ignore = FALSE)),
     "creating tester function test_footests()"
   )
 
@@ -94,6 +99,7 @@ test_that("use_tester works as expected", {
   # Check for unexpected changes to package
   package_state_new <- list.files(".", recursive = TRUE)
   expect_identical(package_state, package_state_new)
+  usethis::proj_set(proj)
 })
 
 
@@ -102,11 +108,12 @@ test_that("use_tester works as expected", {
 test_that("use_test_subdir works as expected", {
   package_state <- list.files(".", recursive = TRUE)
   tpkg <- rprojroot::find_testthat_root_file("testdata", "test_pkg")
+  usethis::proj_set(tpkg)
 
   # Check if tester file is created at the correct path and not empty
   expect_output(
   expect_message(expect_true(
-    use_test_subdir("footests", base_path = tpkg, ignore_tester = FALSE)),
+    use_test_subdir("footests", ignore_tester = FALSE)),
     "creating tester function test_footests()"
   ),
     "Creating"
@@ -128,4 +135,5 @@ test_that("use_test_subdir works as expected", {
   # Check for unexpected changes to package
   package_state_new <- list.files(".", recursive = TRUE)
   expect_identical(package_state, package_state_new)
+  usethis::proj_set(proj)
 })