+++ tests/testthat/test_Test_coverage.R
@@ -27,7 +27,7 @@ test_that("test_Test_coverage works as expected", {
     tested = c(TRUE, FALSE)
   )
 
-  # print.Test_coverage(tdat1)
+  expect_output(print.Test_coverage(tdat1))
   # print.Test_coverage(tdat2)
   # print.Test_coverage(tdat3)
 })