+++ tests/testthat/test_list_test_files.R
@@ -1,35 +1,35 @@
 context("list_test_files")
 
+proj <- usethis::proj_get()
 
 test_that("list_test_files works as expected", {
   #* @testing list_test_files
   #* @testing skip_test_files
 
-  pkg <- devtools::as.package(rprojroot::find_testthat_root_file("testdata", "test_pkg"))
+  tpkg <- rprojroot::find_testthat_root_file("testdata", "test_pkg")
+  usethis::proj_set(tpkg)
 
   # List all test files in /test/test_that
-    tres <- with_mock(
-      `devtools::as.package` = function(...) {pkg},
-      list_test_files()
-    )
+    tres <- list_test_files()
     expect_identical(
       tres,
-      list.files(file.path(pkg$path, "tests", "testthat"), full.names = TRUE)
+      list.files(file.path(tpkg, "tests", "testthat"), full.names = TRUE)
     )
 
 
   # List all test files in /test/test_that, except the ones that contain the
   # testthis tag @skip
-  tres <- with_mock(
-    `devtools::as.package` = function(...) {pkg},
-    list_test_files('.', skip = TRUE)
-  )
+  tres <- list_test_files(skip = TRUE)
 
-  eres <- list.files(file.path(pkg$path, "tests", "testthat"), full.names = TRUE)
+  eres <- list.files(file.path(tpkg, "tests", "testthat"), full.names = TRUE)
   eres <- eres[!grepl('testthis_tags.R', eres, fixed = TRUE)]  # file that contains @skip
 
   expect_identical(
     tres,
     eres
   )
+
+  usethis::proj_set(proj)
 })
+
+usethis::proj_set(proj)